Fixeado la lectura del tamano.

+++ include/sg/stenography/Stenography.h
@@ -13,7 +13,7 @@ class Stenography {
 		virtual deque<char>& extract(vector<char>& host) =0;
 	protected:
 		virtual bool embedable(vector<char>& host, vector<char>& secret) = 0;
-		virtual void embed(char hide, vector<char>& hostArray, u_int& index) = 0;
+		virtual void embed(char hide, vector<char>& hostArray, size_t& index) = 0;
 	private:
 		vector<char>& convertToArray(ifstream& file);
 };
@@ -25,7 +25,7 @@ class LsbStenography : public Stenography {
 		deque<char>& extract(vector<char>& host);
 	protected:
 		bool embedable(vector<char>& host, vector<char>& secret);
-		void embed(char hide, vector<char>& hostArray, u_int& index);
+		void embed(char hide, vector<char>& hostArray, size_t& index);
 	private:
 		int bits;
 };
